Add type to discriminator func encoder setting
This change adds encoder option to set a function that determines
discriminator value from type. In addition it fixes the values of the
constants used in discriminator mode to take 1 and2 instead of 2 and 4.

SetTypeToDiscriminatorFunc allows for customizing the discriminator value for
different types. This may be useful if the golang struct names do not match
the desired values. One example would be if discriminator values in a
protocol require special characters or start with lowercase letter. The
TypeToDiscriminatorFunc implementation may return empty string to suppress
the rendering of discriminator for specific type(s).